requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
black==19.10b0
click==7.1.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.8.0
pylint==2.5.3
python-dotenv==0.13.0
regex==2020.6.8
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
